Reorganize the CQL repo after discussing with team
