+++ package.json
@@ -13,11 +13,11 @@
     "lint:fix": "eslint \"./**/*.js\" --fix",
     "prettier": "prettier --check \"**/*.js\"",
     "prettier:fix": "prettier --write \"**/*.js\"",
-    "test": "./test/run_tests.sh",
-    "test:unit": "jest --testPathPattern=./test",
-    "translate": "node ./scripts/build.js",
-    "test:assurance": "jest --testPathPattern=./assurance",
-    "translate:assurance": "node scripts/build.js assurance/fixtures/cql assurance/fixtures/elm"
+    "test": "./testing-harness/scripts/run_tests.sh",
+    "test:unit": "jest --testPathPattern=test --testPathIgnorePatterns=testing-harness",
+    "translate": "node ./testing-harness/scripts/buildElm.js",
+    "test:testing-harness": "jest --testPathPattern=./testing-harness/test",
+    "translate:testing-harness": "node testing-harness/scripts/buildElm.js testing-harness/test/fixtures/cql testing-harness/test/fixtures/elm"
   },
   "devDependencies": {
     "cql-exec-fhir": "^2.0.0",
