+++ testing-harness/scripts/buildElm.js
@@ -3,14 +3,14 @@ const path = require('path');
 const dotenv = require('dotenv');
 const { Client } = require('cql-translation-service-client');
 
+const cqlPath = process.argv[2] ? path.resolve(process.argv[2]) : path.join(__dirname, '../../cql');
+const buildPath = process.argv[3] ? path.resolve(process.argv[3]) : path.join(__dirname, '../../output-elm');
+
 dotenv.config();
 const TRANSLATION_SERVICE_URL = !process.env.TRANSLATION_SERVICE_URL
   ? 'http://localhost:8080/cql/translator'
   : process.env.TRANSLATION_SERVICE_URL;
 const client = new Client(TRANSLATION_SERVICE_URL);
-
-const cqlPath = process.argv[2] ? path.resolve(process.argv[2]) : path.join(__dirname, '../src');
-const buildPath = process.argv[3] ? path.resolve(process.argv[3]) : path.join(__dirname, '../build');
 
 /**
  * Translate all cql
